Move CMPTxList into seperate file
