+++ src/omnicore/mdex.cpp
@@ -1,6 +1,7 @@
 #include "omnicore/mdex.h"
 
 #include "omnicore/dbtradelist.h"
+#include "omnicore/dbtxlist.h"
 #include "omnicore/errors.h"
 #include "omnicore/fees.h"
 #include "omnicore/log.h"
