+++ src/omnicore/tx.cpp
@@ -6,6 +6,7 @@
 #include "omnicore/dbspinfo.h"
 #include "omnicore/dbstolist.h"
 #include "omnicore/dbtradelist.h"
+#include "omnicore/dbtxlist.h"
 #include "omnicore/dex.h"
 #include "omnicore/fees.h"
 #include "omnicore/log.h"
