+++ src/omnicore/walletfetchtxs.cpp
@@ -8,6 +8,7 @@
 #include "omnicore/walletfetchtxs.h"
 
 #include "omnicore/dbstolist.h"
+#include "omnicore/dbtxlist.h"
 #include "omnicore/log.h"
 #include "omnicore/omnicore.h"
 #include "omnicore/pending.h"
